Update CharacterController.ts
Several small changes for the detection of keyboard keys.
No need for keyCode and fromCharCode

Directly use e.key which allows you to obtain the SPACE, SHIFT and TAB keys and the others (without capital letters). Simpler for users / players.

I put everything in a shitch (e.key) because faster.

All that is set function *** Code (: number) becomes useless.

In addition it seems to me that keyCode and fromCharCode become obsolete or the sound already.